refactor: Section을 추가하고 삭제하는 로직 외부로 분리
